Use CIDR blocks for resource ids
This changes to using the CIDR blocks allocated as the resource ID. This
allows importing resources without having to modify state manually, to
aid in recovery scenarios. As a side effect, changes to the CIDR block
count now require recalculating the entire resource.

Changes to the existing CIDR blocks are ignored, even if they overlap
with the CIDR blocks chosen.

Changes to the pool CIDR blocks are ignored, as long as all of the
selected CIDRs still fall within the pool.