Enable simultaneous allocation of CIDR blocks
The use of the `existing_cidr_block` attributes was a good step in this
direction, but was not compatible with `for_each` or `count` blocks on
resources. Terraform cannot handle resource graph dependencies where
some elements of a `for_each` or `count` resource depend on other
instances of that same resource; these are identified as dependency
cycles.

Instead, the pool and existing allocations are moved to the provider
configuration. This allows resource configurations to remain independent
of each other, yet all resources are allocated from the shared pools.