Update test to better handle concurrency
In some cases, the deletion of the previous state was interfering with
the creation; and the output could contain the same CIDR block twice.
This was due to not having both previously claimed CIDR blocks in the
`claimed_cidr_blocks` list. This update fixes this test construction
issue, and updates the test case to better express the intent; that
simultaneous creation and destruction should be possible, given a
correct configuration.